ai_service/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
psycopg2-binary
sqlalchemy
